Moved Reader and Trainset classes in their own module
