+++ doc/source/dataset.rst
@@ -6,4 +6,3 @@ dataset module
 .. automodule:: surprise.dataset
     :members:
     :exclude-members: BuiltinDataset, read_ratings, DatasetUserFolds,
-        parse_line
