+++ doc/source/reader.rst
@@ -1,0 +1,9 @@
+.. _reader:
+
+Reader class
+============
+
+.. autoclass:: surprise.reader.Reader
+    :members:
+    :exclude-members: parse_line
+
