+++ doc/source/trainset.rst
@@ -1,0 +1,7 @@
+.. _trainset:
+
+Trainset class
+==============
+
+.. autoclass:: surprise.Trainset
+    :members:
